+++ cx-portal/src/components/pages/TechnicalUserManagement/index.tsx
@@ -70,7 +70,7 @@ export default function TechnicalUserManagement() {
             </Button>
           </div>
           <div className="content-technical-user-image">
-            <img src="/edc-connector-text-image.png" alt={'alt tag info'} />
+            <img src="/teaser.png" alt={'alt tag info'} />
           </div>
         </div>